+++ src/FluentDbTools/Abstractions/FluentDbTools.SqlBuilder.Abstractions/IInsertSqlBuilder.cs
@@ -5,6 +5,7 @@ namespace FluentDbTools.SqlBuilder.Abstractions
 {
     public interface IInsertSqlBuilder<TClass> : ISqlFieldBuilder<TClass>
     {
+        IInsertSqlBuilder<TClass> OnTable(string tableName);
         IInsertSqlBuilder<TClass> OnSchema(string schemaName = null, Func<bool> setSchemaNameIfExpressionIsEvaluatedToTrue = null);
         IInsertSqlBuilder<TClass> Fields(Action<IFieldSetterSelector<TClass>> selector);
     }