+++ src/FluentDbTools/Example/Example.FluentDbTools.Migration/MigrationExecutor.cs
@@ -1,24 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-using Example.FluentDbTools.Migration;
+﻿using System.Collections.Generic;
 using FluentDbTools.Common.Abstractions;
 using FluentMigrator.Runner;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace FluentDbTools.Example.Migration
+namespace Example.FluentDbTools.Migration
 {
     public static class MigrationExecutor
     {
-
-        [Obsolete("Please use MigrateUp")]
-        public static void ExecuteMigration(
-            SupportedDatabaseTypes databaseType,
-            Dictionary<string, string> overrideConfig = null)
-        {
-            MigrateUp(databaseType, overrideConfig);
-        }
-
         public static void MigrateUp(
             SupportedDatabaseTypes databaseType,
             Dictionary<string, string> overrideConfig = null)